vlm/requirements.txt
# Python dependencies for VLM inference service.
grpcio>=1.80.0
grpcio-tools>=1.80.0
protobuf>=6.31.1
torch>=2.0.0
transformers>=4.40.0
Pillow>=10.0.0
pyyaml>=6.0
accelerate>=0.26.0
bitsandbytes>=0.42.0
timm>=0.9.0
pytest>=8.0.0
